compile with --use-go-jsonnet
